Handle revoked OAuth tokens when clearing.

+++ app/handlers/checkins.py
@@ -61,12 +61,20 @@ class ClearCheckinsTaskHandler(base.handlers.BaseHandler):
 
     # Clear user data and regenerate it from Foursquare
     user = data.user.User.get_by_foursquare_id(foursquare_id, None)
-    doesnt_want_mail = user.doesnt_want_mail
-    user.delete()
-    user = data.user.User.get_by_foursquare_id(
-        foursquare_id, base.api.Api(oauth_token))
-    user.doesnt_want_mail = doesnt_want_mail
-    user.put()
+    if user:
+      doesnt_want_mail = user.doesnt_want_mail
+      user.delete()
+    else:
+      doesnt_want_mail = None
+    try:
+      user = data.user.User.get_by_foursquare_id(
+          foursquare_id, base.api.Api(oauth_token))
+      user.doesnt_want_mail = doesnt_want_mail
+      user.put()
+    except base.api.ApiException, err:
+      logging.exception(err)
+      logging.warning('Aborting clearing due to an API exception')
+      return
 
     # Clear checkin data and kick off the task to update it
     checkins = data.checkins.Checkins.get_by_foursquare_id(foursquare_id)
